Extract existing limited fetch to reusable method
